Revert "Commit the edit when clicking outside the table."
This reverts commit 152368cb181f34e516d297f34cba7920e501e77c.

It was later discovered that the commit broke the use of combo-boxes.

After some investigation it's because egui renders combo boxes to a different layer, and
thus `resp.clicked_elsewhere` would return true, causing the table to think the user had clicked outside
of the table.

This commit allow comboboxes to work again until a different solution has been found.